feat(#77): Add MariaDB-backup support
Changes:
- Added support of mariadb-backup for mariadb physical backups.
- Deleted obsolete MariaDB Dockerfiles.
- Deleted obsolete documentation.
- Update README.md
