--- .docker/mariadb/Dockerfile-mariadb10_11
@@ -1,16 +1,0 @@
-FROM nixyslab/nxs-backup-bin AS bin
-
-FROM debian:11-slim
-
-RUN apt update \
-    && apt install -yq \
-      lsb-release \
-      curl \
-    && curl -LsS https://r.mariadb.com/downloads/mariadb_repo_setup | bash -s -- --mariadb-server-version="mariadb-10.11" \
-    && apt -yq install mariadb-client \
-    && apt clean \
-    && rm -rf /var/lib/apt/lists/*
-
-COPY --from=bin /nxs-backup /usr/local/bin/nxs-backup
-
-CMD nxs-backup start
