+++ modules/backup/mysql_physical/mysql_physical.go
@@ -1,24 +1,25 @@
-package psql_basebackup
+package mysql_physical
 
 import (
 	"bytes"
 	"errors"
 	"fmt"
-	"net/url"
 	"os"
 	"os/exec"
 	"path"
-	"regexp"
+	"strconv"
 	"strings"
 	"time"
 
 	"github.com/docker/go-units"
 	"github.com/hashicorp/go-multierror"
+	"gopkg.in/ini.v1"
 
-	"github.com/nixys/nxs-backup/ds/psql_connect"
+	"github.com/nixys/nxs-backup/ds/mysql_connect"
 	"github.com/nixys/nxs-backup/interfaces"
 	"github.com/nixys/nxs-backup/misc"
 	"github.com/nixys/nxs-backup/modules/backend/exec_cmd"
+	"github.com/nixys/nxs-backup/modules/backend/files"
 	"github.com/nixys/nxs-backup/modules/backend/targz"
 	"github.com/nixys/nxs-backup/modules/logger"
 	"github.com/nixys/nxs-backup/modules/metrics"
@@ -31,6 +32,7 @@ type job struct {
 	safetyBackup     bool
 	deferredCopying  bool
 	diskRateLimit    int64
+	backupType       misc.BackupType
 	storages         interfaces.Storages
 	targets          map[string]target
 	dumpedObjects    map[string]interfaces.DumpObject
@@ -38,9 +40,12 @@ type job struct {
 }
 
 type target struct {
-	connUrl   *url.URL
-	extraKeys []string
-	gzip      bool
+	extraKeys       []string
+	authFile        *ini.File
+	ignoreDatabases string
+	gzip            bool
+	isSlave         bool
+	prepare         bool
 }
 
 type JobParams struct {
@@ -50,6 +55,7 @@ type JobParams struct {
 	SafetyBackup     bool
 	DeferredCopying  bool
 	DiskRateLimit    int64
+	BackupType       misc.BackupType
 	Storages         interfaces.Storages
 	Sources          []SourceParams
 	Metrics          *metrics.Data
@@ -57,17 +63,30 @@ type JobParams struct {
 
 type SourceParams struct {
 	Name          string
-	ConnectParams psql_connect.Params
+	ConnectParams mysql_connect.Params
+	TargetDBs     []string
+	Excludes      []string
 	ExtraKeys     []string
 	Gzip          bool
 	IsSlave       bool
+	Prepare       bool
+}
+
+func getApp(t misc.BackupType) (app string) {
+	switch t {
+	case misc.MysqlXtrabackup:
+		app = "xtrabackup"
+	case misc.MariadbBackup:
+		app = "mariadb-backup"
+	}
+	return app
 }
 
 func Init(jp JobParams) (interfaces.Job, error) {
 
-	// check if mysqldump available
-	if _, err := exec_cmd.Exec("pg_basebackup", "--version"); err != nil {
-		return nil, fmt.Errorf("Job `%s` init failed. Can't check `pg_basebackup` version. Please install `pg_basebackup`. Error: %s ", jp.Name, err)
+	// check if backup application is available
+	if _, err := exec_cmd.Exec(getApp(jp.BackupType), "--version"); err != nil {
+		return nil, fmt.Errorf("Job `%s` init failed. Can't to check `%s` version. Please install this application. Error: %s ", jp.Name, getApp(jp.BackupType), err)
 	}
 	// check if tar and gzip available
 	if _, err := exec_cmd.Exec("tar", "--version"); err != nil {
@@ -81,13 +100,14 @@ func Init(jp JobParams) (interfaces.Job, error) {
 		safetyBackup:     jp.SafetyBackup,
 		deferredCopying:  jp.DeferredCopying,
 		diskRateLimit:    jp.DiskRateLimit,
+		backupType:       jp.BackupType,
 		storages:         jp.Storages,
 		targets:          make(map[string]target),
 		dumpedObjects:    make(map[string]interfaces.DumpObject),
 		appMetrics: jp.Metrics.RegisterJob(
 			metrics.JobData{
 				JobName:       jp.Name,
-				JobType:       misc.PostgresqlBasebackup,
+				JobType:       jp.BackupType,
 				TargetMetrics: make(map[string]metrics.TargetData),
 			},
 		),
@@ -95,33 +115,27 @@ func Init(jp JobParams) (interfaces.Job, error) {
 
 	for _, src := range jp.Sources {
 
-		for _, key := range src.ExtraKeys {
-			if matched, _ := regexp.MatchString(`(-D|--pgdata=)`, key); matched {
-				return nil, fmt.Errorf("Job `%s` init failed. Forbidden usage \"--pgdata|-D\" parameter as extra_keys for `postgresql_basebackup` jobs type ", jp.Name)
+		_, authFile, err := mysql_connect.GetConnectAndCnfFile(src.ConnectParams, getApp(jp.BackupType))
+		if err != nil {
+			return nil, err
+		}
+
+		var ignoreDBs string
+		if len(src.Excludes) > 0 {
+			ignoreDBs = "--databases-exclude="
+			for _, excl := range src.Excludes {
+				ignoreDBs += excl + " "
 			}
 		}
-
-		cp := src.ConnectParams
-		udb := strings.Split(src.ConnectParams.User, "@")
-		if len(udb) > 1 {
-			cp.Database = udb[1]
-			cp.User = udb[0]
-		}
-
-		connUrl := psql_connect.GetConnUrl(cp)
-		conn, err := psql_connect.GetConnect(connUrl)
-		if err != nil {
-			return nil, fmt.Errorf("Job `%s` init failed. PSQL connect error: %s ", jp.Name, err)
-		}
-		if err = conn.Ping(); err != nil {
-			return nil, fmt.Errorf("Job `%s` init failed. PSQL ping check error: %s ", jp.Name, err)
-		}
-		_ = conn.Close()
+		ignoreDBs = strings.TrimSuffix(ignoreDBs, " ")
 
 		j.targets[src.Name] = target{
-			extraKeys: src.ExtraKeys,
-			gzip:      src.Gzip,
-			connUrl:   connUrl,
+			authFile:        authFile,
+			ignoreDatabases: ignoreDBs,
+			extraKeys:       src.ExtraKeys,
+			gzip:            src.Gzip,
+			isSlave:         src.IsSlave,
+			prepare:         src.Prepare,
 		}
 		j.appMetrics.Job[j.name].TargetMetrics[src.Name] = metrics.TargetData{
 			Source: src.Name,
@@ -148,7 +162,7 @@ func (j *job) GetTempDir() string {
 }
 
 func (j *job) GetType() misc.BackupType {
-	return misc.PostgresqlBasebackup
+	return j.backupType
 }
 
 func (j *job) GetTargetOfsList() (ofsList []string) {
@@ -262,56 +276,108 @@ func (j *job) DoBackup(logCh chan logger.LogRecord, tmpDir string) error {
 	return errs.ErrorOrNil()
 }
 
-func (j *job) createTmpBackup(logCh chan logger.LogRecord, tmpBackupFile, tgtName string, tgt target) error {
+func (j *job) createTmpBackup(logCh chan logger.LogRecord, tmpBackupFile, tgtName string, target target) error {
 
-	var stderr, stdout bytes.Buffer
+	var (
+		stderr, stdout          bytes.Buffer
+		backupArgs, prepareArgs []string
+	)
 
-	tmpBasebackupPath := path.Join(path.Dir(tmpBackupFile), "pg_basebackup_"+tgtName+"_"+misc.GetDateTimeNow(""))
+	tmpBackupPath := path.Join(path.Dir(tmpBackupFile), getApp(j.backupType)+"_"+tgtName+"_"+misc.GetDateTimeNow(""))
 
-	var args []string
-	// define command args
-	// add extra dump cmd options
-	if len(tgt.extraKeys) > 0 {
-		args = append(args, tgt.extraKeys...)
+	authFile, err := files.CreateTmpMysqlAuthFile(target.authFile)
+	if err != nil {
+		logCh <- logger.Log(j.name, "").Errorf("Failed to create tmp auth file. Error: %s", err)
+		return err
 	}
-	// add db connect
-	args = append(args, "--dbname="+tgt.connUrl.String())
-	// add data catalog path
-	args = append(args, "--pgdata="+tmpBasebackupPath)
-	args = append(args, "--format=plain")
-	if j.diskRateLimit > 0 {
-		maxRate := j.diskRateLimit / units.KB
-		if maxRate < 32 {
-			maxRate = 32
-		} else if maxRate > 1024*units.KB {
-			maxRate = 1024 * units.KB
+	defer func() {
+		if err = files.DeleteTmpMysqlAuthFile(authFile); err != nil {
+			logCh <- logger.Log(j.name, "").Errorf("Failed to delete tmp auth file. Error: %s", err)
 		}
-		args = append(args, fmt.Sprintf("--max-rate=%d", maxRate))
+	}()
+
+	// define commands args with auth options
+	backupArgs = append(backupArgs, "--defaults-extra-file="+authFile)
+	prepareArgs = backupArgs
+	// add backup options
+	backupArgs = append(backupArgs, "--backup", "--target-dir="+tmpBackupPath)
+	if target.ignoreDatabases != "" {
+		backupArgs = append(backupArgs, target.ignoreDatabases)
+	}
+	if target.isSlave {
+		backupArgs = append(backupArgs, "--safe-slave-backup")
+	}
+	if j.diskRateLimit != 0 {
+		rateLim := j.diskRateLimit / units.MB
+		if rateLim < 1 {
+			rateLim = 1
+		}
+		// This option limits the number of chunks copied per second. The chunk size is 10 MB.
+		backupArgs = append(backupArgs, "--throttle="+strconv.FormatInt(rateLim, 10))
+	}
+	// add extra backup options
+	if len(target.extraKeys) > 0 {
+		backupArgs = append(backupArgs, target.extraKeys...)
 	}
 
-	cmd := exec.Command("pg_basebackup", args...)
+	cmd := exec.Command(getApp(j.backupType), backupArgs...)
 	cmd.Stdout = &stdout
 	cmd.Stderr = &stderr
 
 	logCh <- logger.Log(j.name, "").Debugf("Dump cmd: %s", cmd.String())
 
-	if err := cmd.Start(); err != nil {
-		logCh <- logger.Log(j.name, "").Errorf("Unable to start pg_basebackup. Error: %s", err)
+	if err = cmd.Start(); err != nil {
+		logCh <- logger.Log(j.name, "").Errorf("Unable to start %s. Error: %v", getApp(j.backupType), err)
 		return err
 	}
-	logCh <- logger.Log(j.name, "").Infof("Starting to dump `%s` source", tgtName)
+	logCh <- logger.Log(j.name, "").Infof("Starting `%s` dump", tgtName)
 
-	if err := cmd.Wait(); err != nil {
-		logCh <- logger.Log(j.name, "").Errorf("Unable to make dump `%s`. Error: %s", tgtName, stderr.String())
+	if err = cmd.Wait(); err != nil {
+		logCh <- logger.Log(j.name, "").Errorf("Unable to dump `%s`. Error: %s", tgtName, err)
+		logCh <- logger.Log(j.name, "").Error(stderr.String())
 		return err
 	}
-	logCh <- logger.Log(j.name, "").Debug("Got psql data. Compressing...")
 
-	if err := targz.Tar(targz.TarOpts{
-		Src:         tmpBasebackupPath,
+	logCh <- logger.Log(j.name, "").Debugf("Exit code: %d", cmd.ProcessState.ExitCode())
+	logCh <- logger.Log(j.name, "").Debugf("STDERR:\n%s", stderr.String())
+
+	if cmd.ProcessState.ExitCode() != 0 {
+		err = j.backupStatusErr(stderr.String())
+		logCh <- logger.Log(j.name, "").Error(err)
+		return err
+	}
+
+	stdout.Reset()
+	stderr.Reset()
+
+	if target.prepare {
+		// add prepare options
+		prepareArgs = append(prepareArgs, "--prepare", "--target-dir="+tmpBackupPath)
+		cmd = exec.Command(getApp(j.backupType), prepareArgs...)
+		cmd.Stdout = &stdout
+		cmd.Stderr = &stderr
+
+		if err = cmd.Run(); err != nil {
+			logCh <- logger.Log(j.name, "").Errorf("Unable to prepare %s. Error: %s", getApp(j.backupType), err)
+			logCh <- logger.Log(j.name, "").Error(stderr.String())
+			return err
+		}
+
+		logCh <- logger.Log(j.name, "").Debugf("Exit code: %d", cmd.ProcessState.ExitCode())
+		logCh <- logger.Log(j.name, "").Debugf("STDERR:\n%s", stderr.String())
+
+		if cmd.ProcessState.ExitCode() != 0 {
+			err = j.backupStatusErr(stderr.String())
+			logCh <- logger.Log(j.name, "").Error(err)
+			return err
+		}
+	}
+
+	if err = targz.Tar(targz.TarOpts{
+		Src:         tmpBackupPath,
 		Dst:         tmpBackupFile,
 		Incremental: false,
-		Gzip:        tgt.gzip,
+		Gzip:        target.gzip,
 		SaveAbsPath: false,
 		RateLim:     j.diskRateLimit,
 		Excludes:    nil,
@@ -323,11 +389,15 @@ func (j *job) createTmpBackup(logCh chan logger.LogRecord, tmpBackupFile, tgtNam
 		}
 		return err
 	}
-	_ = os.RemoveAll(tmpBasebackupPath)
+	_ = os.RemoveAll(tmpBackupPath)
 
-	logCh <- logger.Log(j.name, "").Infof("Dumping of source `%s` completed", tgtName)
+	logCh <- logger.Log(j.name, "").Infof("Dump of `%s` completed", tgtName)
 
 	return nil
+}
+
+func (j *job) backupStatusErr(out string) error {
+	return fmt.Errorf("%s finished not success. Please check result:\n%s", getApp(j.backupType), out)
 }
 
 func (j *job) Close() error {
