+++ modules/backup/psql_logical/psql.go
@@ -1,4 +1,4 @@
-package psql
+package psql_logical
 
 import (
 	"bytes"
